New Version.json
A single place that build tools can access at build time and
the apps can access at runtime. There was no real convention
for this in that both involved a file parse in setup.py. I
just went with the json since it abstracts away from the format
and what not and also is the convention in node's package.json.